Cleaning Up Grate Code (No Collision)
