+++ Dungeon Crab/Assets/Scenes/TTest West.unity
@@ -1964,10 +1964,12 @@ MonoBehaviour:
   endPositionNum: 10
   startPositionNum: 0
   count: 0
-  grateopensPositiveZ: 0
-  grateopensNegativeZ: 0
   grateopensPositiveX: 1
   grateopensNegativeX: 0
+  grateopensPositiveZ: 0
+  grateopensNegativeZ: 0
+  grateopensPositiveY: 0
+  grateopensNegativeY: 0
 --- !u!4 &731648409 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 506371108692209362, guid: 2058db5b2c387a346a041e26955f7a77, type: 3}
@@ -2111,7 +2113,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 765873828}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2dd0578a776cb4cf7aed75e6859674b9, type: 3}
+  m_Script: {fileID: 11500000, guid: dde27b2eb702f4a02b8ad0bc5f7a53eb, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   currentlyOpen: 0
@@ -2120,6 +2122,12 @@ MonoBehaviour:
   endPositionNum: 10
   startPositionNum: 0
   count: 0
+  grateopensPositiveX: 0
+  grateopensNegativeX: 0
+  grateopensPositiveZ: 0
+  grateopensNegativeZ: 0
+  grateopensPositiveY: 1
+  grateopensNegativeY: 0
 --- !u!1001 &780665593
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -5418,10 +5426,12 @@ MonoBehaviour:
   endPositionNum: 10
   startPositionNum: 0
   count: 0
-  grateopensPositiveZ: 0
-  grateopensNegativeZ: 0
   grateopensPositiveX: 1
   grateopensNegativeX: 0
+  grateopensPositiveZ: 0
+  grateopensNegativeZ: 0
+  grateopensPositiveY: 0
+  grateopensNegativeY: 0
 --- !u!4 &2097207784 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 6990225103254478463, guid: 98697f26adf0362478c7ab76e9b1f5ee, type: 3}
