+++ Dungeon Crab/Assets/Scripts/Grate.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9196352120ff44b53af8bfaeaabaa2fd
+guid: dde27b2eb702f4a02b8ad0bc5f7a53eb
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
