--- Dungeon Crab/Assets/Scripts/SideToSideGrate.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dde27b2eb702f4a02b8ad0bc5f7a53eb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
