--- Dungeon Crab/Assets/Scripts/UpDownGrate.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2dd0578a776cb4cf7aed75e6859674b9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
